Dao 120 websocket leakage (#1572)
* refactor and attempt to fix websocket leaks

* fix connection error on socket disconnect web3

* eslint fixed

* update aragon wrapper
