+++ src/App.js
@@ -32,7 +32,7 @@ import {
   DAO_STATUS_LOADING,
   DAO_STATUS_UNLOADED,
 } from './symbols'
-import { useClientWeb3 } from './client-web3'
+import { useClientWeb3 } from './contexts/ClientWeb3Context'
 
 const MIGRATION_BANNER_HIDE = 'MIGRATION_BANNER_HIDE&'
 const MIGRATION_LAST_DATE_ELIGIBLE_TIMESTAMP = new Date(
@@ -456,7 +456,7 @@ export default function AppHooksWrapper(props) {
   const theme = useTheme()
   const clientTheme = useClientTheme()
   const routing = useRouting()
-  const web3 = useClientWeb3()
+  const { web3 } = useClientWeb3()
 
   // analytics
   useEffect(() => {
